requirements.txt
# requirements.txt - WORKING VERSION

# Core
google-generativeai==0.8.3
python-dotenv==1.0.0

# Data Processing
requests==2.31.0

# PDF Processing (we'll add when needed)
# PyPDF2==3.0.1
# pdf2image==1.16.3
# pillow==10.1.0

# Web Framework
flask==3.0.0
flask-cors==4.0.0

# Utils
tqdm==4.66.1


# Core
google-generativeai==0.8.3
python-dotenv==1.0.0

# Data Processing
requests==2.31.0

# Web Framework
flask==3.0.0
flask-cors==4.0.0

# Utils
tqdm==4.66.1

# NEW: For demo UI
jinja2==3.1.2
markupsafe==2.1.3